api/gateway/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
pydantic==2.7.1
aiohttp
python-jose
pydantic-settings
python-multipart